Break apart utils from main module, add some tests
